Move custom preferences Add missing labels
