+++ Quelea/languages/gb.lang
@@ -685,3 +685,7 @@ video.loop.tooltip=Toggle looping on / off
 allow.item.theme.override.global=Allow custom item themes to override global theme
 song.default.theme.label=Song default
 bible.default.theme.label=Bible default
+small.song.position.label=Song information position
+small.bible.position.label=Bible information position
+small.bible.size.label=Bible information size
+small.song.size.label=Song information size
