+++ Quelea/src/main/java/org/quelea/windows/options/customprefs/PercentIntegerSliderControl.java
@@ -1,4 +1,4 @@
-package org.quelea.windows.options;
+package org.quelea.windows.options.customprefs;
 
 import com.dlsc.formsfx.model.structure.IntegerField;
 import com.dlsc.preferencesfx.formsfx.view.controls.SimpleControl;
